requirements.txt
beautifulsoup4
requests
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
pdfminer.six
